requirements.txt
scipy~=1.11.1
numpy~=1.25.2
torchvision~=0.15.2
einops~=0.6.1
pandas~=2.1.0